\name{FontLibrary}
\alias{FontLibrary}
\title{
  Define a Font Library.
}
\description{
  Define a Font Library for querying glyphs.
}
\usage{
FontLibrary(glyphWidth,
            glyphHeight,
            glyphBounds)
}
\arguments{
  \item{glyphWidth, glyphHeight, glyphBounds}{Functions to return glyph
    metric information.}
}
\value{
  A \code{"FontLibrary"} object.
}
\author{
  Paul Murrell
}
\keyword{ dplot }

