\name{grid.latex}
\alias{grid.latex}
\alias{latexGrob}
\alias{author}
\alias{typeset}
\alias{typeset.TeXdocument}
\alias{typeset.TeXfile}
\alias{typeset.character}
\title{
  Render LaTeX in R
}
\description{
  Author, typeset, and render LaTeX in R graphics.
}
\usage{
latexGrob(tex,
          x=0.5, y=0.5, default.units="npc",
          hjust="centre", vjust="centre",
          engine=getOption("latex.engine"),
          fontLib=NULL,
          packages=NULL,
          tinytex=getOption("latex.tinytex"),
          \dots,
          name="LaTeXgrob",
          gp=gpar(),
          vp=NULL) 
grid.latex(\dots)

author(tex,
       engine=getOption("latex.engine"),
       packages=NULL,
       texFile=NULL)

typeset(tex,
        engine=getOption("latex.engine"),
        tinytex=getOption("latex.tinytex"),
        ...)
\method{typeset}{TeXdocument}(tex,
        engine=getOption("latex.engine"),
        tinytex=getOption("latex.tinytex"),
        texFile=NULL,
        ...)
\method{typeset}{TeXfile}(tex,
        engine=getOption("latex.engine"),
        tinytex=getOption("latex.tinytex"),
        ...)
\method{typeset}{character}(tex,
        engine=getOption("latex.engine"),
        tinytex=getOption("latex.tinytex"),
        sig=FALSE, 
        ...)
}
\arguments{
  \item{tex}{LaTeX code.  See Details.}
  \item{x, y}{Numeric values or units specifying where to draw the
    output.}
  \item{default.units}{Units to use if \code{x} or \code{y} are
    numeric.}
  \item{hjust, vjust}{Justification of the output relative to the
    \code{x}/\code{y} location.}
  \item{engine}{The TeX engine that should be used to author, typeset,
    and render the LaTeX.}
  \item{fontLib}{The font libraary the should be used to query fonts
    and glyphs.}
  \item{packages}{The LaTeX packages to be used.}
  \item{tinytex}{Logical indicating whether to use the \pkg{tinytex}
    package (if it is available).}
  \item{name}{Character value giving name for the grob.}
  \item{gp}{Graphical parameter settings.}
  \item{vp}{A viewport or \code{NULL}.}
  \item{texFile}{Name of a file to use for LaTeX code.}
  \item{sig}{Add a signature to the DVI output?}
  \item{\dots}{
    Arguments passed to \code{\link{dviGrob}}.
  }
}
\details{
  \code{author()} can be used to generate a complete LaTeX document
  from a LaTeX fragment, either as a character vector or an external file.

  \code{typeset()} can be used to typeset a LaTeX document, either
  generated by \code{author()} or from an external file.  In the last
  case, it is only appropriate to use \code{sig=TRUE} if the external
  file was generated by this package.
  
  \code{grid.latex()} takes a LaTeX fragment, generates and LaTeX
  document, typesets it, reads the resulting DVI file and renders the
  result.
  
  \code{grid.latex()} and \code{author()}
  expect LaTeX code as a character vector.
  
  \code{typeset()} expects either a \code{"TeXdocument"} or
  a \code{"TeXfile"}, as generated by \code{author}.  If given a
  character value, it treats it as the name of a file containing LaTeX code.
}
\value{
  A \code{"DVIgrob"} object.
}
\author{
  Paul Murrell
}
\seealso{
  \code{\link[dvi]{readDVI}}, 
  \code{\link{TeXengine}},
  \code{\link{LaTeXpackage}}.
}
\keyword{ dplot }

